mcp-tools/github-app-mcp/pyproject.toml
[project]
name = "github-app-mcp"
version = "0.1.0"
description = "An MCP server that performs a strictly allow-listed set of GitHub operations using only a GitHub App identity."
authors = [
  { name = "MCP Generator", email = "mcp@example.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.14"
keywords = ["mcp", "model-context-protocol", "github", "github-app", "security", "audit"]
dependencies = [
  "mcp>=1.0.0",
  "httpx>=0.27.0",
  "PyJWT>=2.9.0",
  "cryptography>=43.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.1.0",
  "ruff>=0.6.0",
]

[project.scripts]
github-app-mcp = "github_app_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/github_app_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "-v",
  "--tb=short",
  "--strict-markers",
  "--disable-warnings",
  "--cov=github_app_mcp",
  "--cov-report=term-missing:skip-covered",
]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
omit = [
  "*/__main__.py",
  "*/server.py",
]

[tool.coverage.report]
fail_under = 95
show_missing = true
skip_covered = true
